added example notebook and refactored imports
